requirements.txt
# requirements.txt
kiteconnect>=5.0.1
pandas>=1.5.0
numpy>=1.23.0
PyYAML>=6.0
requests>=2.28.0
psutil>=5.9.0